feat(Excludes): Add ability to specify excludes
Excludes prevents defaultIncludes from being included and provides a way to remove keys from the transformed output. It can be passed to `createData` as `excludes` or specified with `withExcludes` in the Builder.

BREAKING CHANGE: Excludes has changed the positional order of parameters for `manager.createData` and `new Scope()`.  It comes between `includes` and `identifier`.