fix(Scope): Fix for includes not respecting custom serializers
Previously includes would ignore custom serializers.  This
should now be accomplished in the `scopeData` method.
It would be nice to pull this out to `Scope` at one point, but
right now that messes a lot with the `XmlSerializer`.

BREAKING CHANGE: This changes the interface of the `scopeData` method
in serializers.  It now accepts three arguments — `resource`, `scope`, and `identifier`.
If you have created your own custom serializers, you will need to update
the `scopeData` method on each of them.  If you only used the built-in
serializers, there is nothing you need to do to upgrade.